% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateToxpi.R
\name{calculateToxpi}
\alias{calculateToxpi}
\title{Calculate ToxPi scores}
\usage{
calculateToxpi(dat, mdl)
}
\arguments{
\item{dat}{chemical by assay matrix with appropriate dimnames}

\item{mdl}{list with model specification; see details}
}
\description{
Calculate ToxPi scores
}
\details{
mdl list structure:
\describe{
\item{slices}{named list of character vectors; name gives the slice name
and the character vector gives the assays (columns in dat) to include in
the slice}
\item{weights}{numeric vector of slice weights}
}
}
