% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Download ToxCast Data into zipped folders}
\usage{
downloadData(destDir = getwd())
}
\arguments{
\item{destDir}{the destination directory for the downloaded zip files}
}
\description{
Download ToxCast Data into zipped folders
}
\details{
Downloads two files: (1) ToxCast_20110110.zip containing the
PhI ToxCast data, and (2) INVITRODB_V3_2_SUMMARY.zip containing the most
v3.2 data release (PhIII).
}
