% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalPred.R
\name{evalPred}
\alias{evalPred}
\title{Calculate prediction metrics for the given prediction}
\usage{
evalPred(x, y)
}
\arguments{
\item{x}{prediction (numeric or logical)}

\item{y}{logical classification vector}
}
\description{
Calculate prediction metrics for the given prediction
}
\details{
x and y are subset to complete (neither NA) cases
}
