% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadZipFile.R
\name{freadZipFile}
\alias{freadZipFile}
\title{fread files from zipped archive}
\usage{
freadZipFile(zipDir, zipFl, ...)
}
\arguments{
\item{zipDir}{the zipped folder}

\item{zipFl}{the file within the zipped directory to pass to fread}

\item{...}{passed to \link[data.table:fread]{data.table::fread}}
}
\value{
data.table object
}
\description{
fread files from zipped archive
}
