% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricPlot.R
\name{metricPlot}
\alias{metricPlot}
\title{Make metric plot}
\usage{
metricPlot(metric, invert = FALSE, loadSmryTbl = TRUE, pdfFl = NULL)
}
\arguments{
\item{metric}{character, the metric to plot}

\item{invert}{logical, highest value is darkest when FALSE}

\item{loadSmryTbl}{logical, when TRUE load a clean smryTbl to plot results}

\item{pdfFl}{character, file name/path to save a pdf; do not save to pdf
when NULL}
}
\description{
Make metric plot
}
\details{
Designed for 9in x 6in device w/ 10pt text. Can specify new metrics and
suppress overwriting of standard smryTbl by using \code{loadSmryTbl = FALSE}.
}
