% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankPlot.R
\name{rankPlot}
\alias{rankPlot}
\title{Create ToxPi score rank plots}
\usage{
rankPlot(res, posChems = NULL, negChems = NULL)
}
\arguments{
\item{res}{res matrix from \link{computeToxpi}}

\item{posChems}{vector of codes to color "active" chemcials orange}

\item{negChems}{vector of codes to color "negative" chemicals purple}
}
\description{
Create ToxPi score rank plots
}
